Create Snail
Snail creeps up the vertical pole of height H feet. It goes A feet up per day, and B feet down per night. In which day will the snail reach the top of the pole?

Input data format

On the input the program receives non-negative integers H, A, B, where H > B and A > B. Every integer does not exceed 100.